+++ src/App.jsx
@@ -2,6 +2,7 @@ import React from "react";
 import { Navigation } from "./components/Navigation";
 import { Header } from "./components/Header";
 import { BodyParagraph } from "./components/BodyParagraph";
+import CardInfo from "./components/CardInfo";
 
 import { Footer } from "./components/Footer";
 
@@ -11,7 +12,7 @@ export const App = () => {
       <Navigation />
       <Header />
       <BodyParagraph />
-
+      <CardInfo /> 
       <Footer />
     </div>
   );
